requirements.txt
datasets==3.6.0
pyyaml
requests
tqdm
pillow
numpy
pathlib